fix(jans-config-api): #3725
