+++ jans-config-api/plugins/fido2-plugin/src/main/java/io/jans/configapi/plugin/fido2/service/RegistrationPersistenceService.java
@@ -23,7 +23,5 @@ public class RegistrationPersistenceService extends io.jans.as.common.service.co
     {
     	return userService.getUserInum(userName);
     }
-
-    @Inject
-	private  UserService userService;
+    
 }
